Ef6Repository
Had to remove EF5 from the integration tests since when I added Ef6
nuget package it upgraded Ef5 since it won't let them both be in the
project.  Not sure if there is a way around this so we can have
integration tests for both, but we'll need to look into it.